@@ -11,42 +11,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.13",
-    "@angular/common": "^8.2.13",
-    "@angular/compiler": "^8.2.13",
-    "@angular/core": "^8.2.13",
-    "@angular/forms": "^8.2.13",
-    "@angular/platform-browser": "^8.2.13",
-    "@angular/platform-browser-dynamic": "^8.2.13",
-    "@angular/router": "^8.2.13",
-    "@swimlane/ngx-charts": "^12.0.1",
-    "bootstrap": "^4.3.1",
-    "jquery": "^3.4.1",
-    "ng2-search-filter": "^0.5.1",
-    "popper.js": "^1.16.0",
-    "rxjs": "^6.5.3",
+    "@angular/animations": "^9.1.9",
+    "@angular/cdk": "^9.2.4",
+    "@angular/common": "^9.1.9",
+    "@angular/compiler": "^9.1.9",
+    "@angular/core": "^9.1.9",
+    "@angular/forms": "^9.1.9",
+    "@angular/platform-browser": "^9.1.9",
+    "@angular/platform-browser-dynamic": "^9.1.9",
+    "@angular/router": "^9.1.9",
+    "@swimlane/ngx-charts": "^14.0.0",
+    "@swimlane/ngx-datatable": "^17.0.0",
+    "bootstrap": "^4.5.0",
+    "jquery": "^3.5.1",
+    "ngx-filter-pipe": "^2.1.2",
+    "popper.js": "^1.16.1",
+    "rxjs": "^6.5.5",
     "tslib": "^1.10.0",
-    "zone.js": "^0.10.2"
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.17",
-    "@angular/cli": "^8.3.26",
-    "@angular/compiler-cli": "^8.2.13",
-    "@angular/language-service": "^8.2.13",
-    "@types/node": "^12.12.6",
-    "@types/jasmine": "^3.4.6",
+    "@angular-devkit/build-angular": "^0.901.7",
+    "@angular/cli": "^9.1.7",
+    "@angular/compiler-cli": "^9.1.9",
+    "@angular/language-service": "^9.1.9",
+    "@types/jasmine": "^3.5.10",
     "@types/jasminewd2": "^2.0.8",
-    "codelyzer": "^5.2.0",
+    "@types/node": "^14.0.5",
+    "codelyzer": "^5.2.2",
     "jasmine-core": "^3.5.0",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^4.4.1",
+    "jasmine-spec-reporter": "^5.0.2",
+    "karma": "^5.0.9",
     "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage-istanbul-reporter": "^2.1.0",
-    "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "protractor": "^5.4.2",
-    "ts-node": "^8.4.1",
-    "tslint": "^5.20.1",
-    "typescript": "~3.5.3"
+    "karma-coverage-istanbul-reporter": "^3.0.2",
+    "karma-jasmine": "^3.2.0",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "protractor": "^7.0.0",
+    "ts-node": "^8.10.1",
+    "tslint": "^6.1.2",
+    "typescript": "^3.8.3"
   }
 }
