@@ -1,15 +1,11 @@
-from django.conf.urls import url
-from django.urls import path, include
-
 from storage import views
+from django.urls import path
 from rest_framework import routers
 
-router = routers.DefaultRouter()
-router.register(r'graphs', views.GraphViewSet)
+router = routers.SimpleRouter()
+router.register('api/graphs', views.GraphViewSet)
 
 urlpatterns = [
-    url('^$', views.HomePageView.as_view()),
-    url('^graph/$', views.Graph.as_view())
+    path('', views.HomePageView.as_view())
 ]
-
 urlpatterns += router.urls
