@@ -1,7 +1,5 @@
 from rest_framework import views
-from rest_framework import status
 from rest_framework import viewsets
-from rest_framework.utils import json
 from rest_framework import serializers
 from rest_framework.response import Response
 from rest_framework.renderers import TemplateHTMLRenderer
@@ -11,23 +9,15 @@ from storage import models
 
 
 class HomePageView(views.APIView):
+    """
+    APIView for rendering angular front
+    """
     renderer_classes = [TemplateHTMLRenderer]
     template_name = 'index.html'
 
     @staticmethod
     def get(request):
         return Response()
-
-
-class Graph(views.APIView):
-    @staticmethod
-    def get(request):
-        graph = models.Graph.objects.get(id=request.query_params.get('id'))
-        return Response(GraphSerializer(graph).data)
-
-    @staticmethod
-    def post(request):
-        return Response('saved')
 
 
 class DataSerializer(serializers.ModelSerializer):
@@ -37,6 +27,7 @@ class DataSerializer(serializers.ModelSerializer):
 
 
 class GraphSerializer(serializers.ModelSerializer):
+
     data = DataSerializer(many=True)
 
     class Meta:
@@ -45,5 +36,8 @@ class GraphSerializer(serializers.ModelSerializer):
 
 
 class GraphViewSet(viewsets.ModelViewSet):
-    queryset = models.Graph.objects.all()
+    """
+    CRUD for graphs
+    """
     serializer_class = GraphSerializer
+    queryset = models.Graph.objects.all()
